chore: refactor events
