+++ frontend/src/container/InfraMonitoringK8s/Clusters/ClusterDetails/ClusterDetails.tsx
@@ -12,6 +12,7 @@ import {
 	initialQueryState,
 } from 'constants/queryBuilder';
 import ROUTES from 'constants/routes';
+import { K8sCategory } from 'container/InfraMonitoringK8s/constants';
 import { QUERY_KEYS } from 'container/InfraMonitoringK8s/EntityDetailsUtils/utils';
 import {
 	CustomTimeType,
@@ -43,8 +44,8 @@ import {
 import { GlobalReducer } from 'types/reducer/globalTime';
 import { v4 as uuidv4 } from 'uuid';
 
+import ClusterEvents from '../../EntityDetailsUtils/EntityEvents';
 import { ClusterDetailsProps } from './ClusterDetails.interfaces';
-import ClusterEvents from './Events';
 import ClusterLogs from './Logs';
 import ClusterMetrics from './Metrics';
 import ClusterTraces from './Traces';
@@ -527,6 +528,8 @@ function ClusterDetails({
 							isModalTimeSelection={isModalTimeSelection}
 							handleTimeChange={handleTimeChange}
 							selectedInterval={selectedInterval}
+							category={K8sCategory.CLUSTERS}
+							queryKey="clusterEvents"
 						/>
 					)}
 				</>
