--- frontend/src/container/InfraMonitoringK8s/Clusters/ClusterDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import ClusterEvents from './ClusterEvents';
-
-export default ClusterEvents;
