--- frontend/src/container/InfraMonitoringK8s/Deployments/DeploymentDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import DeploymentEvents from './DeploymentEvents';
-
-export default DeploymentEvents;
