--- frontend/src/container/InfraMonitoringK8s/Namespaces/NamespaceDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import NamespaceEvents from './NamespaceEvents';
-
-export default NamespaceEvents;
