--- frontend/src/container/InfraMonitoringK8s/Nodes/NodeDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import NodeEvents from './NodeEvents';
-
-export default NodeEvents;
